I NEEDED TO USE PARSEFLOAT THE WHOLE DAMN TIME AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA

+++ script.js
@@ -324,9 +324,9 @@ function pythagGenSolve() {
 
 function expSolve() {
   var expA = document.getElementById("expA").value;
-  var expB = parseInt(document.getElementById("expB").value);
+  var expB = parseFloat(document.getElementById("expB").value);
   var expN = document.getElementById("expN").value;
-  var expResult = expA * Math.pow((1 + parseInt(expB)), expN);
+  var expResult = expA * Math.pow((1 + parseFloat(expB)), expN);
   
   document.getElementById("expResult").innerHTML = expResult;
 }
